requirements.txt
photoscript
langchain
langchain-community
langchain-core
pillow

